requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
networkx>=3.0.0
gradio-client>=0.7.0
python-docx>=0.8.11
PyPDF2>=3.0.0
tiktoken>=0.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
python-louvain>=0.16
requests>=2.31.0
python-dotenv>=1.0.0
setuptools>=65.0.0
hnswlib>=0.7.0
scipy>=1.10.0
llama-parse>=0.4.0
nest-asyncio>=1.5.0
streamlit>=1.28.0
pyvis>=0.3.2
openai>=1.0.0
# Database for NodeRAG storage - migrated from NeonDB to Neo4j
neo4j~=5.28.0
# Additional dependencies for API service
pandas>=2.1.0
aiohttp>=3.8.0
tqdm>=4.65.0